Add files via upload
The sales analysis on Superstore dataset is a comprehensive study that aims to analyze the sales performance of a fictional retail company called “Superstore”. The dataset used in this analysis contains information about sales transactions, customers, products, and geographical locations. The analysis involves using Pivot table to provide insights into the sales performance of Superstore. The data is cleaned and transformed using Power Query . The analysis explores trends in sales over time, product categories, segments, and geographical regions. The results of the analysis provide valuable insights into Superstore’s sales performance, highlighting areas of strength and weakness. The analysis reveals that Superstore’s sales are strongest in the East and West regions of the United States, and that technology is the most profitable product category. The analysis also identifies opportunities for improvement, such as increasing sales in the Central region and improving customer retention. Overall, the sales analysis on Superstore dataset provides a valuable tool for businesses to better understand their sales performance and make data-driven decisions to improve their sales. The Superstore dataset is collected from a reliable source,(Kaggle ).The  Visualization part of the superstore dataset. In united state the best generating revenue state is from California.
Under sub-categories phones are the top performing product while fastener product are underperforming, Chairs are not doing bad as a product ,also I will advise marketing to focus more on Office supplies and more of furniture to increase sales.
The consumer segment has the largest customer base while corporate and Home office has the smallest customer , I will suggest that Superstore should increase their marketing strategy so has to be an increase in the store growth .
Sales in west region has the highest revenue followed by the East region . This two regions bring in more revenue on like south region and Central region , still on the fact that Superstore should focus more on marketing optimization on these two regions so as to bring in more Revenue to the company.